chore(server-split): v0.4.x PR-A.1 — auth/upload/admin-user routes moved to routes/auth.py (13 endpoints + 13 Pydantic models)
docs/design/v0.4.x-server-split.md 8-PR sequence 중 두 번째 PR.
PR-A.0 foundation 위에 실제 endpoint 이동을 수행. trust-boundary
이동이라 audit_endpoint_paths.py + pytest 모두 zero-diff 확인 후 land.

## 이동된 surface

### 13 endpoints (724 lines 서버에서 제거)
- /login/, /signup/, /upload/  (auth + 파일 업로드, 286 lines /upload/ 포함)
- /admin/users, /admin/users/approve, /admin/users/reject, /admin/users/deactivate
- /password/change, /password/reset/confirm
- /admin/users/issue-reset-token
- /api-keys/issue, /api-keys/list, /api-keys/revoke

### 13 Pydantic 모델 (auth 표면 응집도 유지)
- LoginRequest, LoginResponse, SignupRequest, SignupResponse
- UploadResponse
- PasswordChangeRequest
- AdminUserApproveRequest, AdminUserRejectRequest, AdminUserDeactivateRequest
- AdminIssueResetTokenRequest, PasswordResetConfirmRequest
- ApiKeyIssueRequest, ApiKeyRevokeRequest

### 모듈-level 상수
- _SIGNUP_ACCEPTED_MSG  (signup 응답 메시지, enumeration defense)

## 변경 패턴 (handler 본문 byte-identical)

각 handler 본문은 100% 보존. 두 가지만 바꿈:
1. `@app.<m>` → `@router.<m>`  (single source for the 13 decorators)
2. /upload/ 한정으로 def 시그니처 직후 3-line 의존성 주입 shim:
       rag_engine = get_rag_engine()
       file_processor = get_file_processor()
   (다른 12개 handler는 module-level singleton 비사용 — shim 불필요)

## server_llmwiki.py 변경

- 13 endpoint + 13 Pydantic + _SIGNUP_ACCEPTED_MSG 제거 (~720 lines)
- `from routes.auth import router as auth_router` + `app.include_router(auth_router)` 와이어
- include_router 위치: set_rate_limiter() 직후 (DI singleton 등록 후)
- 14 unused import 자동 제거 (authenticate / signup / approve_user / reject_user / deactivate_user / list_users / change_password / issue_reset_token / consume_reset_token / RESET_TOKEN_TTL_SEC / issue_api_key / revoke_api_key / list_api_keys / uuid)
- `resolve_api_key_principal` re-import 추가 `# noqa: F401`  (tests/test_api_key_middleware.py 가 `server_llmwiki.resolve_api_key_principal` 으로 접근 — back-compat 위해 re-export)

## tests/test_video_asr.py 갱신

`UploadAcceptsVideoExtensionsTests` 가 server_llmwiki.py 만 스캔했으나
/upload/ handler 가 routes/auth.py 로 이동. setUpClass 가 양쪽 모두
스캔하도록 변경 (다음 server-split PR 이 이 invariant 를 깨도 검증
가능).

## 검증 (design memo §5 contract)

| invariant | 결과 |
|---|---|
| 1. URL byte-identical | ✓ `python scripts/audit_endpoint_paths.py origin/main` → **OK: 125 endpoints identical to origin/main** |
| 2. Middleware order | ✓ `test_middleware_stack_includes_rate_limit_and_no_cache` pass |
| 3. _write_audit emit | ✓ 24 사이트 모두 동일 위치 + sqlite3 INSERT (routes/_helpers SOT) |
| 4. /static mount | ✓ `test_static_mount_present` pass |
| 5. startup event | ✓ `test_startup_event_registered` pass |
| 6. pytest 회귀 | ✓ **3294 passed, 0 failed** (2 .env-flakes deselected — pre-existing, 본 PR 무관) |
| ruff F-class | ✓ All checks passed |

## 변경 통계

```
 routes/auth.py          | 806 ++++++++++++++++++++++++++++++++++++++++++++++++
 server_llmwiki.py       | 731 +------------------------------------------
 tests/test_video_asr.py |  12 +-
 3 files changed, 825 insertions(+), 724 deletions(-)
```

server_llmwiki.py: 5894 → 5170 lines (-724). 8-PR roadmap 목표 (최종
~500 lines) 진행률 ≈ 12% (724/5894 제거).

## Out of scope (다음 PR 들로)

- PR-B: /llm/* + /admin/llm/* routes (Ollama / 모델 매핑)
- PR-C: /jobs/* (W8 workspace jobs)
- PR-D: /artifacts/* + /admin/artifacts/* (W7-A data artifacts)
- PR-E: Phase 7 자기진화 + 멀티모달
- PR-F: /code/* (코딩 에이전트)
- PR-G: 잔여 /admin/*
- PR-H: server_llmwiki.py 최종 정리 (~500 lines target)

## Quality Delta vs baseline

`Quality delta: exempt (label: chore)`. URL 표면 byte-identical, `core/` 무관.
